LayoutTests/storage/domstorage/localstorage/blocked-file-access-permitted-by-quirk-expected.txt

CONSOLE MESSAGE: line 8: PASS: window.localStorage WAS accessible

Test that the local storage quirk allows us to access localStorage from a file URL, even if unversal access is turned off.
